requirements.txt
python-dotenv
discord
giphy_client
